# src/main/webapp/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.1",
    "title": "BBVA GCP Web Services",
    "contact": {
      "name": "API Support",
      "email": "paul.rivera.tadeo.contractor@dev.bbva.com"
    }
  },
...
  "paths": {
    "/v1/board/pagination": {
      "post": {
        "tags": [
          "Board"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/boardPaginationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/boardPaginationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/board/list": {
      "post": {
        "tags": [
          "Board"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/boardListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/boardListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/bui/pagination": {
      "post": {
        "tags": [
          "Bui"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/buiPaginationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/buiPaginationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bui/readonly": {
      "post": {
        "tags": [
          "Bui"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/buiReadonlyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/buiReadonlyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/source/pagination": {
      "post": {
        "tags": [
          "Source"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sourcePaginationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sourcePaginationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/source/readonly": {
      "post": {
        "tags": [
          "Source"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sourceReadonlyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sourceReadonlyResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/authentication/validate": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "validate email GCP.",
        "requestBody": {
          "description": "data user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "google": {
                          "type": "string"
                        },
                        "fullName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "photoUrl": {
                          "type": "string"
                        }
                      }
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "role": {
                            "type": "string"
                          },
                          "menus": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer"
                                },
                                "label": {
                                  "type": "string"
                                },
                                "icon": {
                                  "type": "string"
                                },
                                "url": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/issue_ticket/insert": {
      "post": {
        "tags": [
          "issue_ticket"
        ],
        "summary": "generate issue ticket.",
        "requestBody": {
          "description": "data issue",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workOrderId": {
                    "type": "integer"
                  },
                  "workOrderCode": {
                    "type": "string"
                  },
                  "folio": {
                    "type": "string"
                  },
                  "boardId": {
                    "type": "integer"
                  },
                  "projectId": {
                    "type": "integer"
                  },
                  "sourceId": {
                    "type": "string"
                  },
                  "sourceName": {
                    "type": "string"
                  },
                  "flowType": {
                    "type": "integer"
                  },
                  "registerUserId": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  },
                  "workOrderDetail": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "templateId": {
                          "type": "integer"
                        },
                        "issueCode": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/issue_ticket/listSourcesGenerated": {
      "post": {
        "tags": [
          "issue_ticket"
        ],
        "summary": "list sources generated.",
        "requestBody": {
          "description": "data sources",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sourceTicketRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sourceTicketResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/issue_ticket/listIssuesGenerated": {
      "post": {
        "tags": [
          "issue_ticket"
        ],
        "summary": "list issues jira generated.",
        "requestBody": {
          "description": "data issues",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sourceTicketRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sourceTicketResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/authentication/validate/permissions": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "validate permissions rol GCP.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDtoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateDtoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rol/detail/permissions": {
      "post": {
        "tags": [
          "rol"
        ],
        "summary": "list permissions of rol.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolDataDetailPermissionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolDataDetailPermissionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rol/register": {
      "post": {
        "tags": [
          "rol"
        ],
        "summary": "register rol.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolDataResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rol/update": {
      "post": {
        "tags": [
          "rol"
        ],
        "summary": "update rol.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolDataResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "boardPaginationRequest": {
        "type": "object",
        "properties": {
          "records_amount": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "id_fuente": {
            "type": "integer"
          },
          "fuente_origen": {
            "type": "string"
          },
          "fuente_datio": {
            "type": "string"
          },
          "funcional_descripcion": {
            "type": "string"
          }
        }
      },
      "boardPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "count": {
                "type": "integer"
              },
              "pages_amount": {
                "type": "integer"
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "fuente": {
                      "type": "integer"
                    },
                    "deuda_tecnica": {
                      "type": "integer"
                    },
                    "fuente_antigua": {
                      "type": "string"
                    },
                    "proyecto": {
                      "type": "integer"
                    },
                    "proyecto_sdatool": {
                      "type": "string"
                    },
                    "proyecto_nombre": {
                      "type": "string"
                    },
                    "fuente_origen": {
                      "type": "string"
                    },
                    "estado_fuente": {
                      "type": "integer"
                    },
                    "estado_fuente_descripcion": {
                      "type": "string"
                    },
                    "estado_deuda": {
                      "type": "integer"
                    },
                    "estado_deuda_descripcion": {
                      "type": "string"
                    },
                    "tipologia": {
                      "type": "integer"
                    },
                    "tipologia_descripcion": {
                      "type": "string"
                    },
                    "tabla_master": {
                      "type": "string"
                    },
                    "estado_deuda_tecnica": {
                      "type": "integer"
                    },
                    "estado_deuda_tecnica_descripcion": {
                      "type": "string"
                    },
                    "tabla_raw": {
                      "type": "string"
                    },
                    "complejidad_deuda": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "boardListRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
      "boardListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "code": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "buiPaginationRequest": {
        "type": "object",
        "properties": {
          "records_amount": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "sdatool": {
            "type": "string"
          },
          "proposed_table": {
            "type": "string"
          },
          "analyst_in_charge": {
            "type": "string"
          },
          "folio_code": {
            "type": "string"
          },
          "id_fuente": {
            "type": "integer"
          },
          "tipo": {
            "type": "integer"
          },
          "estado": {
            "type": "integer"
          }
        }
      },
      "buiPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "count": {
                "type": "integer"
              },
              "pages_amount": {
                "type": "integer"
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "sourceId": {
                      "type": "integer"
                    },
                    "fuenteAnteriorId": {
                      "type": "integer"
                    },
                    "sourceName": {
                      "type": "string"
                    },
                    "folioId": {
                      "type": "integer"
                    },
                    "folioCode": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "integer"
                    },
                    "sdatool": {
                      "type": "string"
                    },
                    "projectName": {
                      "type": "string"
                    },
                    "resolutionSourceTypeId": {
                      "type": "integer"
                    },
                    "resolutionSourceType": {
                      "type": "string"
                    },
                    "ingestSourceTypeId": {
                      "type": "integer"
                    },
                    "ingestSourceType": {
                      "type": "string"
                    },
                    "statusFolioTypeId": {
                      "type": "integer"
                    },
                    "statusFolioType": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "buiReadonlyRequest": {
        "type": "object",
        "properties": {
          "buiId": {
            "type": "integer"
          }
        }
      },
      "buiReadonlyResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "roadmap": {
                "type": "object",
                "properties": {
                  "folio": {
                    "type": "string"
                  },
                  "project": {
                    "type": "string"
                  },
                  "sdatool": {
                    "type": "string"
                  },
                  "tabla": {
                    "type": "string"
                  },
                  "historia": {
                    "type": "string"
                  },
                  "periocidad": {
                    "type": "string"
                  },
                  "fechaRegistro": {
                    "type": "string"
                  },
                  "descTabla": {
                    "type": "string"
                  }
                }
              },
              "dictamen": {
                "type": "object",
                "properties": {
                  "estatus": {
                    "type": "string"
                  },
                  "analista_proyecto": {
                    "type": "string"
                  },
                  "analista_ca": {
                    "type": "string"
                  },
                  "resolucion": {
                    "type": "string"
                  },
                  "folio_reuso": {
                    "type": "string"
                  },
                  "fecha_sprint_cierre": {
                    "type": "string"
                  },
                  "comentario_resolucion": {
                    "type": "string"
                  },
                  "tipo": {
                    "type": "string"
                  }
                }
              },
              "tdsDictaminada": {
                "type": "object",
                "properties": {
                  "id_tds": {
                    "type": "string"
                  },
                  "fuenteId": {
                    "type": "string"
                  },
                  "fuente_tds": {
                    "type": "string"
                  },
                  "historia_ingestar": {
                    "type": "string"
                  },
                  "comentario": {
                    "type": "string"
                  },
                  "estado_tds": {
                    "type": "string"
                  },
                  "id_reemplazado": {
                    "type": "string"
                  },
                  "estado_deuda": {
                    "type": "string"
                  },
                  "comentario_deuda": {
                    "type": "string"
                  }
                }
              },
              "ingesta": {
                "type": "object",
                "properties": {
                  "persistencia_destino": {
                    "type": "string"
                  },
                  "dashboard": {
                    "type": "string"
                  },
                  "fecha_dashboard": {
                    "type": "string"
                  },
                  "comentario_general": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "sourcePaginationRequest": {
        "type": "object",
        "properties": {
          "records_amount": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "id_fuente": {
            "type": "integer"
          },
          "nombre_fuente": {
            "type": "string"
          },
          "uuaa_aplicativo": {
            "type": "string"
          },
          "estado": {
            "type": "integer"
          },
          "origen": {
            "type": "integer"
          },
          "estado_deuda": {
            "type": "integer"
          },
          "uuaa_master": {
            "type": "string"
          },
          "descripcion_tds": {
            "type": "string"
          },
          "tabla_master": {
            "type": "string"
          },
          "propietario_global": {
            "type": "integer"
          }
        }
      },
      "sourcePaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "count": {
                "type": "integer"
              },
              "pages_amount": {
                "type": "integer"
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "sourceId": {
                      "type": "integer"
                    },
                    "id": {
                      "type": "string"
                    },
                    "fuente": {
                      "type": "string"
                    },
                    "descripcion": {
                      "type": "string"
                    },
                    "estado_id": {
                      "type": "integer"
                    },
                    "estado": {
                      "type": "string"
                    },
                    "tipo_origen_id": {
                      "type": "integer"
                    },
                    "tipo_origen": {
                      "type": "string"
                    },
                    "uuaa_master_id": {
                      "type": "integer"
                    },
                    "uuaa_master": {
                      "type": "string"
                    },
                    "deuda_id": {
                      "type": "integer"
                    },
                    "deuda": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "sourceReadonlyRequest": {
        "type": "object",
        "properties": {
          "sourceId": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}